added changes related to review comments
